backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
PyMySQL==1.1.0
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Validation
WTForms==3.0.1
email-validator==2.0.0

# Development
Flask-Debugtoolbar==0.14.1
